+++ src/lit-element.ts
@@ -123,8 +123,8 @@ export class LitElement extends UpdatingElement {
     // (3) shadowRoot.adoptedStyleSheets polyfilled: append styles after
     // rendering
     if (window.ShadyCSS !== undefined && !window.ShadyCSS.nativeShadow) {
-      window.ShadyCSS.ScopingShim.prepareAdoptedCssText(styles.map((s) => s.cssText),
-                                            this.localName);
+      window.ShadyCSS.ScopingShim.prepareAdoptedCssText(
+          styles.map((s) => s.cssText), this.localName);
     } else if (supportsAdoptingStyleSheets) {
       (this.renderRoot as ShadowRoot).adoptedStyleSheets =
           styles.map((s) => s.styleSheet!);